fix: sleep now
