--- %ProgramData%/Microsoft/Windows/UUS/State/_active.uusver
@@ -1,1 +1,0 @@
-923.330.111.0
