+++ src/constants/firebase/index.js
@@ -1,5 +1,5 @@
 export const FIREBASE_REALTIME_DB_URL =
-  "";
-export const FIREBASE_API_KEY = "";
+  "https://sleep-calculator-f9192-default-rtdb.firebaseio.com/";
+export const FIREBASE_API_KEY = "AIzaSyCx1IY4AKjClEoSuWCIhuhebkcIln1bGac";
 export const FIREBASE_AUTH_SIGN_IN_URL = `https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=${FIREBASE_API_KEY}`;
 export const FIREBASE_AUTH_SIGN_UP_URL = `https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=${FIREBASE_API_KEY}`;
